fix: リレーションシップを持ったモデルを集約
